+++ Cargo.toml
@@ -1,39 +1,10 @@
-[package]
-name = "wasm-ipld"
-version = "0.1.0"
-authors = ["Adin Schmahmann <adin.schmahmann@gmail.com>"]
-edition = "2018"
+[workspace]
 
-[lib]
-crate-type = ["cdylib", "rlib"]
-
-[features]
-default = ["wee_alloc"]
-
-[dependencies]
-#wasm-bindgen = "0.2.63"
-
-# The `console_error_panic_hook` crate provides better debugging of panics by
-# logging them with `console.error`. This is great for development, but requires
-# all the `std::fmt` and `std::panicking` infrastructure, so isn't great for
-# code size when deploying.
-#console_error_panic_hook = { version = "0.1.6", optional = true }
-
-# `wee_alloc` is a tiny allocator for wasm that is only ~1K in code size
-# compared to the default allocator's ~10K. It is slower than the default
-# allocator, however.
-#
-# Unfortunately, `wee_alloc` requires nightly Rust when targeting wasm for now.
-wee_alloc = { version = "0.4.5", optional = true }
-serde_ipld_dagcbor = {version = "0.1.2"}
-serde-transcode = {version = "1.1.1"}
-bt_bencode = {version = "0.6.1"}
-libipld = {version = "0.13.1"}
-serde = {version = "1.0.136"}
-integer-encoding = { version = "3.0.3"}
-num_enum = { version = "0.5.7" }
-
-hex = { version = "0.4" }
+members = [
+    "bencode",
+    "bt_filev1",
+    "wac"
+]
 
 [profile.release]
 # Tell `rustc` to optimize for small code size.
